# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ParseInterestRequest {
  /**
   * @minLength 2
   * @maxLength 1000
   */
  rawText: string;
  userId?: string;
}

export type InterestSpecDataIntentType =
  (typeof InterestSpecDataIntentType)[keyof typeof InterestSpecDataIntentType];

export const InterestSpecDataIntentType = {
  monitor: "monitor",
  alert: "alert",
  opportunity: "opportunity",
  match: "match",
  creator_watch: "creator_watch",
  travel: "travel",
  local_signal: "local_signal",
} as const;

export type InterestSpecDataUrgency =
  (typeof InterestSpecDataUrgency)[keyof typeof InterestSpecDataUrgency];

export const InterestSpecDataUrgency = {
  high: "high",
  medium: "medium",
  low: "low",
} as const;

export type InterestSpecDataTrustNeed =
  (typeof InterestSpecDataTrustNeed)[keyof typeof InterestSpecDataTrustNeed];

export const InterestSpecDataTrustNeed = {
  high: "high",
  medium: "medium",
  low: "low",
} as const;

export type InterestSpecDataMatchMode =
  | (typeof InterestSpecDataMatchMode)[keyof typeof InterestSpecDataMatchMode]
  | null;

export const InterestSpecDataMatchMode = {
  companion: "companion",
  friend: "friend",
  collaborate: "collaborate",
  meal_mate: "meal_mate",
  date: "date",
} as const;

export type InterestSpecDataPrivacyLevel =
  (typeof InterestSpecDataPrivacyLevel)[keyof typeof InterestSpecDataPrivacyLevel];

export const InterestSpecDataPrivacyLevel = {
  public: "public",
  friends: "friends",
  private: "private",
} as const;

export type InterestSpecDataSuggestedSourcesItem =
  (typeof InterestSpecDataSuggestedSourcesItem)[keyof typeof InterestSpecDataSuggestedSourcesItem];

export const InterestSpecDataSuggestedSourcesItem = {
  youtube: "youtube",
  twitter: "twitter",
  reddit: "reddit",
  rss: "rss",
  match: "match",
} as const;

export interface InterestSpecData {
  intentType: InterestSpecDataIntentType;
  topic: string;
  entities: string[];
  locationScope?: string | null;
  urgency: InterestSpecDataUrgency;
  desiredOutcome: string;
  trustNeed: InterestSpecDataTrustNeed;
  matchMode?: InterestSpecDataMatchMode;
  privacyLevel: InterestSpecDataPrivacyLevel;
  negativeConstraints?: string[];
  suggestedSources: InterestSpecDataSuggestedSourcesItem[];
}

export type AgentStepStatus =
  (typeof AgentStepStatus)[keyof typeof AgentStepStatus];

export const AgentStepStatus = {
  success: "success",
  partial: "partial",
  failed: "failed",
} as const;

export interface AgentStep {
  agent: string;
  status: AgentStepStatus;
  message: string;
  durationMs?: number;
}

export interface ParsedInterestResult {
  spec: InterestSpecData;
  steps: AgentStep[];
}

export interface GenerateAlertsRequest {
  spec: InterestSpecData;
  /**
   * @minimum 1
   * @maximum 5
   */
  count?: number;
}

export type AlertDataFreshness =
  (typeof AlertDataFreshness)[keyof typeof AlertDataFreshness];

export const AlertDataFreshness = {
  live: "live",
  hot: "hot",
  recent: "recent",
  older: "older",
} as const;

export type AlertDataSourceType =
  (typeof AlertDataSourceType)[keyof typeof AlertDataSourceType];

export const AlertDataSourceType = {
  youtube: "youtube",
  twitter: "twitter",
  reddit: "reddit",
  rss: "rss",
  match: "match",
} as const;

export type AlertDataSource = {
  type: AlertDataSourceType;
  name: string;
  /** Direct link to the original source. Opens externally when tapped. */
  url?: string;
};

export interface AlertData {
  title: string;
  summary: string;
  reason: string;
  /**
   * @minimum 0
   * @maximum 100
   */
  confidence: number;
  freshness: AlertDataFreshness;
  source: AlertDataSource;
  tags: string[];
  /** @minimum 0 */
  minutesAgo?: number;
}

export interface GeneratedAlertsResult {
  alerts: AlertData[];
  steps: AgentStep[];
}

export type RegisterDeviceRequestPlatform =
  (typeof RegisterDeviceRequestPlatform)[keyof typeof RegisterDeviceRequestPlatform];

export const RegisterDeviceRequestPlatform = {
  ios: "ios",
  android: "android",
  web: "web",
} as const;

export interface RegisterDeviceRequest {
  /**
   * @minLength 4
   * @maxLength 128
   */
  deviceId: string;
  /**
   * @minLength 4
   * @maxLength 256
   */
  expoPushToken: string;
  platform: RegisterDeviceRequestPlatform;
}

export interface RegisterDeviceResult {
  ok: boolean;
}

export interface TrackInterestRequest {
  /**
   * @minLength 1
   * @maxLength 128
   */
  interestId: string;
  /**
   * @minLength 4
   * @maxLength 128
   */
  deviceId: string;
  spec: InterestSpecData;
  /** @maxLength 1000 */
  rawText?: string;
}

export interface TrackInterestResult {
  ok: boolean;
  interestId: string;
}

export interface PushTestRequest {
  /**
   * @minLength 4
   * @maxLength 128
   */
  deviceId: string;
}

export interface PushTestResult {
  ok: boolean;
  ticket: string;
}

export type UntrackInterest200 = {
  ok: boolean;
};